move functions to functions.py
